Add pre-commit hooks
Add a few pre-commit hooks to ensure code quality is maintaneid
througout the development.

yapf is one of the most used code-formatters for python, it will
automatically format the code if it's not up to its standards.

The custom pre-commit script is going to run pylint on the code dir
after every commit. This linter checks wether the code is PEP-8
compliant or not.

Also added hooks to check for trailing whitespaces, missing end of
files, for invalid yaml and a checker to see if large files were added.